+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Added
 
+- Collection of relationships between Administrators and Roles
+- Collection of Role entities
 - Collection of Computer entities
 - Collection of API Key entities
 - Collection of Administrator entities